+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [8.4.0] - 2023-09-30
+
+### Added
+
+- transformer config facility
+- xml white space remove utility
+- StreamHelper.resolveReader() method
+
 ### Changed
 
 - fj-bom version set to 1.4.7